Update: convert the string of modules returned from API into a struct
The API returns the "Modules" field as a string. Previously this was left as is,
but presented problems in clients who needed to manually parse those string. Now,
it has been converted to a struct with fields Name and Version. Any client parsing
will just be for the name, which is easily accessible now workspace.Modules[*].Name.