Add: private_registry.go - contains function to retrieve info about the latest modules in the private registry
private_registry contains one function at this point. It is designed to pull the latest version of all modules
and return that to the end user. As of the writing of the function PrivateRegistryModules(), the latest version
of a module is at index[0] of the API result, and this is hard coded, no checking is done yet at this point, but
will be implemented in the future in case the API return changes.

Also updated buildUrl to explicilty be buildExplorerUrl to allow for this new endpoint and its equivalanet
function buildRegistryUrl